# CHERRY-PICK: chore_: 'CheckAllChannelsPermissions' fn optimization