'CheckAllChannelsPermissions' fn optimization

Instead of checking permissions for each channel, do a permissions check first and store the results in channels

Tested on status community, where we have 18 channel permissions and 116 channels
The average result of calling CheckAllChannelsPermissions

Before
2 accounts: ~2m 10s
1 account: ~1m 10s
After:
2 accounts ~25s
1 account: ~12s

Note: The results are average and can vary based on the PC and network speed

Closes # https://github.com/status-im/status-go/issues/5179

Cherry-pick: https://github.com/status-im/status-go/pull/5196