@status-im-auto 2024-05-23
### Jenkins Builds
| :grey_question: | Commit | :hash: | Finished (UTC) | Duration | Platform | Result |
|-|-|-|-|-|-|-|
| :heavy_check_mark: | 9883ef54e5f90faef4d0855ad766acd7bc5348de | [#1](https://ci.status.im/job/status-go/job/prs/job/linux/job/PR-5219/1/) | 2024-05-23 16:33:31 | ~4 min | `linux` | [:package:`zip`](https://status-im-prs.ams3.cdn.digitaloceanspaces.com/status-go-desktop-240523-162849-9883ef54e5f90faef4d0855ad766acd7bc5348de-pr5219.zip) |